- Additional changes for publish
