+++ WereViewApp/App_Start/AppConfig.cs
@@ -1,14 +1,11 @@
 ﻿using WereViewApp.Models.Context;
 using WereViewApp.Models.POCO.IdentityCustomization;
-using WereViewApp.Modules.Mail;
 using WereViewApp.Modules.Session;
 using WereViewApp.Modules.TimeZone;
 using System;
-using System.Configuration;
 using System.Linq;
 using System.Reflection;
 using System.Web;
-using System.Web.Mvc;
 using DevMvcComponent;
 using DevMvcComponent.Error;
 using DevMvcComponent.Mail;
@@ -78,6 +75,7 @@ namespace WereViewApp {
                 }
                 return _setting;
             }
+            private set { _setting = value; }
         }
 
         /// <summary>
@@ -145,20 +143,14 @@ namespace WereViewApp {
 
             using (var db = new DevIdentityDbContext()) {
                 CreateDefaultCoreSetting();
-
                 _setting = db.CoreSettings.FirstOrDefault();
+                Setting = db.CoreSettings.FirstOrDefault();
                 if (_setting == null) {
                     throw new Exception("Couldn't create or get the core settings. Please check the creation.");
                 }
                 Zone.LoadTimeZonesIntoMemory();
-
-                AppVar.IsInTestEnvironment = Setting.IsInTestingEnvironment;
-                AppVar.Name = Setting.ApplicationName.ToString();
-                AppVar.Subtitle = Setting.ApplicationSubtitle.ToString();
-                AppVar.Setting = Setting;
-
+                AppVar.Setting = _setting;
                 AppVar.SetCommonMetaDescriptionToEmpty();
-
                 SetupDevMvcComponent();
                 //if false then no email on error.
                 Config.IsNotifyDeveloper = Setting.NotifyDeveloperOnError;
@@ -187,148 +179,6 @@ namespace WereViewApp {
         public static void SetGlobalError(ErrorCollector error) {
             HttpContext.Current.Session[SessionNames.Error] = error;
         }
-
-    }
-    /// <summary>
-    /// Application Global Variables
-    /// </summary>
-    public struct AppVar {
-
-        #region Enums
-
-        #endregion
-
-        #region Constants
-
-
-
-
-        #endregion
-
-        #region Connection Strings and Constants
-        //public const string DefaultConnection = @"Data Source=(LocalDb)\v11.0;AttachDbFilename=|DataDirectory|\WereViewApp-Accounts.mdf;Initial Catalog=WereViewApp-Accounts;Integrated Security=True";
-        private static readonly string DefaultConnection = ConfigurationManager
-                              .ConnectionStrings["DefaultConnection"]
-                              .ConnectionString;
-        public enum ConnectionStringType {
-            DefaultConnection,
-            Secondary
-        };
-
-
-        #endregion
-
-        #region Propertise
-        static string _productNameMeta;
-        /// <summary>
-        /// Application Name
-        /// </summary>
-        public static string Name;
-        /// <summary>
-        /// Application Subtitle
-        /// </summary>
-        public static string Subtitle;
-        /// <summary>
-        /// Is application in testing environment or not?
-        /// </summary>
-        public static bool IsInTestEnvironment;
-
-
-
-        public static CoreSetting Setting;
-        /// <summary>
-        /// Get the application URL based on the application environment.
-        /// Without slash.
-        /// </summary>
-        public static string Url {
-            get {
-                if (IsInTestEnvironment) {
-                    return AppConfig.Setting.TestingUrl;
-                }
-                return AppConfig.Setting.LiveUrl;
-            }
-        }
-
-        public static MailSender Mailer = new MailSender();
-        #endregion
-
-        #region Functions
-
-
-
-        public static string GetConnectionString(ConnectionStringType type) {
-            switch (type) {
-                case ConnectionStringType.DefaultConnection:
-                    return DefaultConnection;
-                case ConnectionStringType.Secondary:
-                    break;
-                default:
-                    break;
-            }
-            return null;
-        }
-        static string GetCommonMetadescription() {
-            string finalMeta = "";
-            if (_productNameMeta == null) {
-                var nameList = Name.Split(' ').ToList();
-                nameList.Add(Name);
-                nameList.Add(Subtitle);
-                foreach (var item in nameList) {
-                    if (finalMeta.Equals("")) {
-                        finalMeta += ",";
-                    }
-                    finalMeta += item;
-                }
-                _productNameMeta = finalMeta;
-            }
-            return _productNameMeta;
-        }
-        internal static void SetCommonMetaDescriptionToEmpty() {
-            _productNameMeta = null;
-        }
-
-        public static ActionResult GetFriendlyError(string title, string message) {
-            var dictionary = new ViewDataDictionary(){
-              {"Title",title},
-              {"ErrorMessage",message}
-            };
-            return new ViewResult() {
-                ViewName = "_FriendlyError",
-                ViewData = dictionary
-            };
-        }
-
-        public static ActionResult GetAuthenticationError(string title, string message) {
-            var dictionary = new ViewDataDictionary(){
-              {"Title", title},
-              {"ErrorMessage",message}
-            };
-            return new ViewResult() {
-                ViewName = "_AuthenticationError",
-                ViewData = dictionary
-            };
-        }
-
-        public static void GetTitlePageMeta(dynamic viewBag, string title, string msg = "", string meta = null, string keywords = null) {
-            viewBag.Title = title;
-            viewBag.Message = msg;
-            viewBag.Meta = meta + "," + GetCommonMetadescription();
-            viewBag.Keywords = keywords + "," + GetCommonMetadescription();
-        }
-        public static void SetSavedStatus(dynamic viewBag, string msg = null) {
-            if (msg == null) {
-                msg = "Your previous transaction is successfully saved.";
-            }
-            viewBag.Success = msg;
-        }
-
-        public static void SetErrorStatus(dynamic viewBag, string msg = null) {
-            if (msg == null) {
-                msg = "Your last transaction is not saved.";
-            }
-            viewBag.ErrorSave = msg;
-        }
-        #endregion
 
     }
 }
