+++ WereViewApp/Areas/Admin/Views/Shared/_NavBar.cshtml
@@ -21,7 +21,6 @@
                         <a href="#">Developers Organsim <b class="caret"></b></a>
                         <ul class="dropdown animated fadeIn">
                             @Html.GetMenu("Admin")
-                            <li>@Html.ActionLink("Feedback", "Index", new { Controller = "Feedback" })</li>
                         </ul>
                     </li>
                     <li>
