+++ WereViewApp/Views/Shared/_LoginPartial.cshtml
@@ -25,7 +25,7 @@
     </li>
 } else {
     <li title="Login in WeReviewApp.com">
-        <a href="@AppVar.Url/Login" title="Login in WeReviewApp.com">
+        <a href="/Login" title="Login in WeReviewApp.com">
             <h1 title="Login in WeReviewApp.com">
                 <strong title="Login in WeReviewApp.com">
                     <em title="Login in WeReviewApp.com">
@@ -36,7 +36,7 @@
         </a>
         <ul class="animated dropdown fadeIn">
             <li>
-                <a href="@AppVar.Url/Login" title="Login in WeReviewApp.com">
+                <a href="/Login" title="Login in WeReviewApp.com">
                     <h1 title="Login in WeReviewApp.com">
                         <strong title="Login in WeReviewApp.com">
                             <em title="Login in WeReviewApp.com">
@@ -47,7 +47,7 @@
                 </a>
             </li>
             <li>
-                <a href="@AppVar.Url/Register" title="Register in WeReviewApp.com">
+                <a href="/Register" title="Register in WeReviewApp.com">
                     <h1 title="Register in WeReviewApp.com">
                         <strong title="Register in WeReviewApp.com">
                             <em title="Register in WeReviewApp.com">
