+++ WereViewApp/Web.config
@@ -39,13 +39,13 @@
   </appSettings>
 
   <system.web>
-    <customErrors mode="off" />
+    <!--<customErrors mode="Off" />-->
     
     <!-- cookie free domain -->
     <sessionState cookieless="UseDeviceProfile" regenerateExpiredSessionId="true" timeout="50" />
     <!--<sessionState cookie less="UseUri"/>-->
     <caching>
-      <outputCache enableOutputCache="false" />
+      <outputCache enableOutputCache="true" />
       <outputCacheSettings>
         <outputCacheProfiles>
           <add name="YearNoParam" duration="31104000" varyByParam="none" varyByCustom="none" location="Server" />
@@ -103,8 +103,8 @@
     </modules>
     <validation validateIntegratedModeConfiguration="false" />
     
-    <httpErrors errorMode="Detailed" />
-    <!--<httpErrors errorMode="Custom" existingResponse="Replace" defaultResponseMode="ExecuteURL">
+    <!--<httpErrors errorMode="Detailed" />-->
+    <httpErrors errorMode="Custom" existingResponse="Replace" defaultResponseMode="ExecuteURL">
       <remove statusCode="400" subStatusCode="-1" />
       <remove statusCode="403" subStatusCode="-1" />
       <remove statusCode="404" subStatusCode="-1" />
@@ -125,7 +125,7 @@
       <error statusCode="451" responseMode="ExecuteURL" path="/Errors/Error_451" />
       <error statusCode="500" responseMode="ExecuteURL" path="/Errors/Error_500" />
       <error statusCode="502" responseMode="ExecuteURL" path="/Errors/Error_502" />
-    </httpErrors>-->             
+    </httpErrors>             
     <asp scriptErrorSentToBrowser="true" />
     <httpProtocol>
       <customHeaders>
